requirements.txt
Django==5.1.4
openai==1.98.0
google-generativeai==0.8.5
requests==2.32.4
python-dotenv==1.0.0
Pillow==10.0.1 